Adds point_generator_test.py.
Point generator uses a generator function to make the x and y values for each point.
This means I don't need to worry about creating a 2d matrix beforehand.
Results from the pool of workers are aggregated in a 1d matrix, which is then reshaped.
Pyplot shows the mandlebrot cardiod, so I'm counting this as a success.