Adds make_char_mandlebrot function to cbor.py. Changes picture.py to use it by default.
Previously picture.py was making a numpy matrix containing float64 values. This is useful
when you can expect a decimal value as a result, like in the log normalized version of the program.
But you don't need to normalize anything if you reduce the max iterations. Turns out 100 is
enough to get a good image, and pyplot can make a heatmap out of integers. Because of this,
the program can store values as unsigned chars and cut the matrix's memory use to 1/8 the value.
This is the first time it ran at step 0.0001 without killing the process, and the reduced iterations
didnt appear to affect the picture quality.